permissions update
added cros-platform support for requesting runtime permissions.
this was put off because the actual runtime request function is API 23+ and we're targetting 16+.
this is handled by only requesting on 23+ and just showing an alert box otherwise (if permissions are missing).

additionally, initialization is performed after permissions result, so we don't have to restart the app.