added text field and remove specific control
added a text field control that you can edit with a popup dialog.
currently it just sends an event to the server any time you change the content.
next, i'll be adding ways to request and edit the content from the server side.

added a way to remove a specific control now that they all are guaranteed to have unique ids.