battery optimization and privacy
previously, we had both sensors and networking run even when the app was in the background, allowing the screen to be turned off and still use it as a sensor.
however, this was using a lot of power if the user doesn't actually close the app, which a lot of people tend to not do.
additionally, if they don't actually close it, it'll never disconnect from the server (unless they actually lose internet connection).
this commit changes all sensor data to no longer run in the background and also puts the networking thread to sleep.