fix crashing bug from rotation on some devices
some devices (not my testing one) were crashing upon rotation, which mine wasn't doing for some reason.
regardless, rotation was not intended because it would make using the custom interface hard to use with acceleration/gyroscope being used too (changing constantly).
and it would totally mess up the aspect ratio of the custom control canvas, which is bad cause the controls are specified in relative size.