gui update
i want to add more controls and options in the future, but they're already cluttering up the display, which takes space from the custom controls canvas.
thus, i've moved everything we currently have into a navigation view so that it's out of the way when not in use.
now we just have a toolbar at the top of the screen to open it, and otherwise we have an edge-to-edge custom control canvas.