landscape controls and font size
added an option for controls to be landscape.
this affects the rendering of individual controls; the app itself is always in portrait mode.
this was added to make landscape text for e.g. game controllers.
also added font size, though not all controls accept a custom value for it yet (buttons do).

changed the joystick code so that we send at most ten messages per second while moving it around.
this was becoming an issue for the NetsBlox server with only a single phone; just imagine a whole classroom...