checkbox update and bug fix
updated the checkbox/toggleswitch rendering logic to be identical to the iOS version.
added the landscape and readonly settings and implemented the fontSize placeholder we had for it.

fixed a bug where we were returning the opposite check state from the ckeckblox click event.

added the setToggleState function to IToggleable (like in iOS).
similarly, buttons are now a separate IPushable interface with its own network request.